yolo

+++ include/SimpleEstimator.h
@@ -17,7 +17,7 @@ private:
 
     //[first][sec][{noOut, noPaths, noI}]
     //uint32_t *precal;
-    uint16_t *precal;
+    uint32_t *precal;
     uint32_t precal_L2;
 
     inline uint32_t precal_noOut(uint32_t L1, uint32_t L2);
